+++ src/services/oidc.security.service.ts
@@ -1,7 +1,7 @@
-import { HttpParams } from '@angular/common/http';
+import { HttpParams, HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable, NgZone } from '@angular/core';
 import { Router } from '@angular/router';
-import { BehaviorSubject, from, Observable, Subject, throwError as observableThrowError, timer } from 'rxjs';
+import { BehaviorSubject, from, Observable, Subject, throwError as observableThrowError, timer, of } from 'rxjs';
 import { catchError, filter, map, race, shareReplay, switchMap, switchMapTo, take, tap } from 'rxjs/operators';
 import { OidcDataService } from '../data-services/oidc-data.service';
 import { AuthWellKnownEndpoints } from '../models/auth.well-known-endpoints';
@@ -70,7 +70,8 @@ export class OidcSecurityService {
         private oidcSecurityValidation: OidcSecurityValidation,
         private tokenHelperService: TokenHelperService,
         private loggerService: LoggerService,
-        private zone: NgZone
+        private zone: NgZone,
+        private readonly httpClient: HttpClient
     ) {
         this.onModuleSetup.pipe(take(1)).subscribe(() => {
             this.moduleSetup = true;
@@ -236,6 +237,7 @@ export class OidcSecurityService {
         this.oidcSecurityCommon.customRequestParams = params;
     }
 
+    // Code Flow with PCKE or Implicit Flow
     authorize(urlHandler?: (url: string) => any) {
         if (this.authWellKnownEndpoints) {
             this.authWellKnownEndpointsLoaded = true;
@@ -253,7 +255,7 @@ export class OidcSecurityService {
 
         this.resetAuthorizationData(false);
 
-        this.loggerService.logDebug('BEGIN Authorize, no auth data');
+        this.loggerService.logDebug('BEGIN Authorize Code Flow, no auth data');
 
         let state = this.oidcSecurityCommon.authStateControl;
         if (!state) {
@@ -265,40 +267,126 @@ export class OidcSecurityService {
         this.oidcSecurityCommon.authNonce = nonce;
         this.loggerService.logDebug('AuthorizedController created. local state: ' + this.oidcSecurityCommon.authStateControl);
 
-        if (this.authWellKnownEndpoints) {
-            const url = this.createAuthorizeUrl(
-                this.authConfiguration.redirect_url,
-                nonce,
-                state,
-                this.authWellKnownEndpoints.authorization_endpoint
-            );
+        let url = '';
+        // Code Flow
+        if (this.authConfiguration.response_type === 'code') {
 
-            if (urlHandler) {
-                urlHandler(url);
+            // code_challenge with "S256"
+            const code_verifier = 'C' + Math.random() + '' + Date.now() + '' + Date.now() + Math.random();
+            const code_challenge = this.oidcSecurityValidation.generate_code_verifier(code_verifier);
+
+            this.oidcSecurityCommon.code_verifier = code_verifier;
+
+            if (this.authWellKnownEndpoints) {
+                url = this.createAuthorizeUrl(true, code_challenge,
+                    this.authConfiguration.redirect_url,
+                    nonce,
+                    state,
+                    this.authWellKnownEndpoints.authorization_endpoint
+                );
             } else {
-                this.redirectTo(url);
+                this.loggerService.logError('authWellKnownEndpoints is undefined');
             }
+        } else { // Implicit Flow
+
+            if (this.authWellKnownEndpoints) {
+                url = this.createAuthorizeUrl(false, '',
+                    this.authConfiguration.redirect_url,
+                    nonce,
+                    state,
+                    this.authWellKnownEndpoints.authorization_endpoint
+                );
+            } else {
+                this.loggerService.logError('authWellKnownEndpoints is undefined');
+            }
+        }
+
+        if (urlHandler) {
+            urlHandler(url);
         } else {
-            this.loggerService.logError('authWellKnownEndpoints is undefined');
+            this.redirectTo(url);
         }
     }
 
-    authorizedCallback(hash?: string) {
+    // Code Flow
+    authorizedCallbackWithCode(urlToCheck: string) {
+        const urlParts = urlToCheck.split('?');
+        const params = new HttpParams({
+            fromString: urlParts[1]
+        });
+        const code = params.get('code');
+        const state = params.get('state');
+        const session_state = params.get('session_state');
+
+        if (code && state && session_state) {
+            this.requestTokensWithCode(code, state, session_state);
+        }
+    }
+
+    // Code Flow
+    requestTokensWithCode(code: string, state: string, session_state: string | null) {
         this._isModuleSetup
             .pipe(
                 filter((isModuleSetup: boolean) => isModuleSetup),
                 take(1)
             )
             .subscribe(() => {
-                this.authorizedCallbackProcedure(hash);
+                this.requestTokensWithCodeProcedure(code, state, session_state);
             });
     }
 
-    private redirectTo(url: string) {
-        window.location.href = url;
+    // Code Flow with PCKE
+    requestTokensWithCodeProcedure(code: string, state: string, session_state: string | null) {
+        let tokenRequestUrl = '';
+        if (this.authWellKnownEndpoints && this.authWellKnownEndpoints.token_endpoint) {
+            tokenRequestUrl = `${this.authWellKnownEndpoints.token_endpoint}`;
+        }
+
+        // TODO validate state early instead of waiting for the callback with the tokens
+
+        let headers: HttpHeaders = new HttpHeaders();
+        headers = headers.set('Content-Type', 'application/x-www-form-urlencoded');
+
+        let data = `grant_type=authorization_code&client_id=${this.authConfiguration.client_id}`
+            + `&code_verifier=${this.oidcSecurityCommon.code_verifier}&code=${code}&redirect_uri=${this.authConfiguration.redirect_url}`;
+        if (this.oidcSecurityCommon.silentRenewRunning === 'running') {
+            data = `grant_type=authorization_code&client_id=${this.authConfiguration.client_id}`
+                + `&code_verifier=${this.oidcSecurityCommon.code_verifier}&code=${code}&redirect_uri=${this.authConfiguration.silent_redirect_url}`;
+        }
+
+        this.httpClient
+            .post(tokenRequestUrl, data, { headers: headers })
+            .pipe(
+            map(response => {
+                    let obj: any = new Object;
+                    obj = response;
+                    obj.state = state;
+                    obj.session_state = session_state;
+
+                    this.authorizedCodeFlowCallbackProcedure(obj);
+                }),
+            catchError(error => {
+                    this.loggerService.logError(error);
+                    this.loggerService.logError(`OidcService code request ${this.authConfiguration.stsServer}`);
+                    return of(false);
+                })
+            )
+            .subscribe();
     }
 
-    private authorizedCallbackProcedure(hash?: string) {
+    // Code Flow
+    private authorizedCodeFlowCallbackProcedure(result: any) {
+        const silentRenew = this.oidcSecurityCommon.silentRenewRunning;
+        const isRenewProcess = silentRenew === 'running';
+
+        this.loggerService.logDebug('BEGIN authorized Code Flow Callback, no auth data');
+        this.resetAuthorizationData(isRenewProcess);
+
+        this.authorizedCallbackProcedure(result, isRenewProcess);
+    }
+
+    // Implicit Flow
+    private authorizedImplicitFlowCallbackProcedure(hash?: string) {
         const silentRenew = this.oidcSecurityCommon.silentRenewRunning;
         const isRenewProcess = silentRenew === 'running';
 
@@ -307,12 +395,33 @@ export class OidcSecurityService {
 
         hash = hash || window.location.hash.substr(1);
 
-        const result: any = hash.split('&').reduce(function(resultData: any, item: string) {
+        const result: any = hash.split('&').reduce(function (resultData: any, item: string) {
             const parts = item.split('=');
             resultData[<string>parts.shift()] = parts.join('=');
             return resultData;
         }, {});
 
+        this.authorizedCallbackProcedure(result, isRenewProcess);
+    }
+
+    // Implicit Flow
+    authorizedImplicitFlowCallback(hash?: string) {
+        this._isModuleSetup
+            .pipe(
+                filter((isModuleSetup: boolean) => isModuleSetup),
+                take(1)
+            )
+            .subscribe(() => {
+                this.authorizedImplicitFlowCallbackProcedure(hash);
+            });
+    }
+
+    private redirectTo(url: string) {
+        window.location.href = url;
+    }
+
+    // Implicit Flow
+    private authorizedCallbackProcedure(result: any, isRenewProcess: boolean) {
         this.oidcSecurityCommon.authResult = result;
 
         // reset the history to remove the tokens
@@ -409,14 +518,14 @@ export class OidcSecurityService {
 
         return new Observable<boolean>(observer => {
             // flow id_token token
-            if (this.authConfiguration.response_type === 'id_token token') {
+            if (this.authConfiguration.response_type === 'id_token token' || this.authConfiguration.response_type === 'code') {
                 if (isRenewProcess && this._userData.value) {
                     this.oidcSecurityCommon.sessionState = result.session_state;
                     observer.next(true);
                     observer.complete();
                 } else {
                     this.oidcSecurityUserService.initUserData().subscribe(() => {
-                        this.loggerService.logDebug('authorizedCallback id_token token flow');
+                        this.loggerService.logDebug('authorizedCallback (id_token token || code) flow');
 
                         const userData = this.oidcSecurityUserService.getUserData();
 
@@ -504,16 +613,39 @@ export class OidcSecurityService {
         this.loggerService.logDebug('RefreshSession created. adding myautostate: ' + this.oidcSecurityCommon.authStateControl);
 
         let url = '';
-        if (this.authWellKnownEndpoints) {
-            url = this.createAuthorizeUrl(
-                this.authConfiguration.silent_redirect_url,
-                nonce,
-                state,
-                this.authWellKnownEndpoints.authorization_endpoint,
-                'none'
-            );
+
+        // Code Flow
+        if (this.authConfiguration.response_type === 'code') {
+
+            // code_challenge with "S256"
+            const code_verifier = 'C' + Math.random() + '' + Date.now() + '' + Date.now() + Math.random();
+            const code_challenge = this.oidcSecurityValidation.generate_code_verifier(code_verifier);
+
+            this.oidcSecurityCommon.code_verifier = code_verifier;
+
+            if (this.authWellKnownEndpoints) {
+                url = this.createAuthorizeUrl(true, code_challenge,
+                    this.authConfiguration.silent_redirect_url,
+                    nonce,
+                    state,
+                    this.authWellKnownEndpoints.authorization_endpoint,
+                    'none'
+                );
+            } else {
+                this.loggerService.logWarning('authWellKnownEndpoints is undefined');
+            }
         } else {
-            this.loggerService.logWarning('authWellKnownEndpoints is undefined');
+            if (this.authWellKnownEndpoints) {
+                url = this.createAuthorizeUrl(false, '',
+                    this.authConfiguration.silent_redirect_url,
+                    nonce,
+                    state,
+                    this.authWellKnownEndpoints.authorization_endpoint,
+                    'none'
+                );
+            } else {
+                this.loggerService.logWarning('authWellKnownEndpoints is undefined');
+            }
         }
 
         this.oidcSecurityCommon.silentRenewRunning = 'running';
@@ -607,7 +739,7 @@ export class OidcSecurityService {
         this.oidcSecurityCommon.isAuthorized = true;
     }
 
-    private createAuthorizeUrl(redirect_url: string, nonce: string, state: string, authorization_endpoint: string, prompt?: string): string {
+    private createAuthorizeUrl(isCodeFlow: boolean, code_challenge: string, redirect_url: string, nonce: string, state: string, authorization_endpoint: string, prompt?: string): string {
         const urlParts = authorization_endpoint.split('?');
         const authorizationUrl = urlParts[0];
         let params = new HttpParams({
@@ -620,6 +752,12 @@ export class OidcSecurityService {
         params = params.append('scope', this.authConfiguration.scope);
         params = params.append('nonce', nonce);
         params = params.append('state', state);
+
+        if (isCodeFlow) {
+
+            params = params.append('code_challenge', code_challenge);
+            params = params.append('code_challenge_method', 'S256');
+        }
 
         if (prompt) {
             params = params.append('prompt', prompt);
@@ -733,6 +871,27 @@ export class OidcSecurityService {
 
     private silentRenewEventHandler(e: CustomEvent) {
         this.loggerService.logDebug('silentRenewEventHandler');
-        this.authorizedCallback(e.detail);
+
+        if (this.authConfiguration.response_type === 'code') {
+
+            const urlParts = e.detail.toString().split('?');
+            const params = new HttpParams({
+                fromString: urlParts[1]
+            });
+            const code = params.get('code');
+            const state = params.get('state');
+            const session_state = params.get('session_state');
+            const error = params.get('error');
+            if (code && state && session_state) {
+                this.requestTokensWithCodeProcedure(code, state, session_state);
+            }
+            if (error) {
+                this.loggerService.logDebug(e.detail.toString());
+            }
+
+        } else {
+            // ImplicitFlow
+            this.authorizedImplicitFlowCallback(e.detail);
+        }
     }
 }