+++ tests/services/oidc.security.service.spec.ts
@@ -77,6 +77,7 @@ describe('OidcSecurityService', () => {
         oidcSecurityService.setupModule(openIDImplicitFlowConfiguration);
 
         let value = oidcSecurityService.createAuthorizeUrl(
+		    false, '', // Implicit Flow
             openIDImplicitFlowConfiguration.redirect_url,
             'nonce',
             'state',
@@ -113,6 +114,7 @@ describe('OidcSecurityService', () => {
         oidcSecurityService.authConfiguration.init(openIDImplicitFlowConfiguration);
 
         let value = oidcSecurityService.createAuthorizeUrl(
+		    false, '', // Implicit Flow
             openIDImplicitFlowConfiguration.redirect_url,
             'nonce',
             'state',
@@ -184,6 +186,7 @@ describe('OidcSecurityService', () => {
         });
 
         let value = oidcSecurityService.createAuthorizeUrl(
+		    false, '', // Implicit Flow
             openIDImplicitFlowConfiguration.redirect_url,
             'nonce',
             'state',
@@ -225,6 +228,7 @@ describe('OidcSecurityService', () => {
         });
 
         let value = oidcSecurityService.createAuthorizeUrl(
+		    false, '', // Implicit Flow
             openIDImplicitFlowConfiguration.redirect_url,
             'nonce',
             'state',
@@ -266,7 +270,7 @@ describe('OidcSecurityService', () => {
         expect(value).toEqual(expectValue);
     });
 
-    it('authorizedCallback should correctly parse hash params', () => {
+    it('authorizedImplicitFlowCallback should correctly parse hash params', () => {
         spyOn(oidcSecurityService, 'getSigningKeys').and.returnValue(empty());
 
         const resultSetter = spyOnProperty(
@@ -282,7 +286,7 @@ describe('OidcSecurityService', () => {
             state: 'testState',
         };
 
-        (oidcSecurityService as OidcSecurityService).authorizedCallback(hash);
+        (oidcSecurityService as OidcSecurityService).authorizedImplicitFlowCallback(hash);
 
         expect(resultSetter).not.toHaveBeenCalled();
 
@@ -295,7 +299,7 @@ describe('OidcSecurityService', () => {
         expectedResult['access_token'] = 'ACCESS-TOKEN==';
         expectedResult['state'] = 'test=State';
 
-        (oidcSecurityService as OidcSecurityService).authorizedCallback(hash);
+        (oidcSecurityService as OidcSecurityService).authorizedImplicitFlowCallback(hash);
         expect(resultSetter).toHaveBeenCalledWith(expectedResult);
     });